+++ R/reassemble_detection.R
@@ -1,7 +1,7 @@
 #' @title Reassemble detections from clips
 #'
 #' @description \code{reassemble_detection} reassembles detections made on clips so they refer to the original sound files
-#' @param detection Data frame or selection table (using the warbleR package's format, see \code{\link[warbleR]{selection_table}}) containing the start and end of the signals. Must contained at least the following columns: "sound.files", "selec", "start" and "end". 
+#' @param detection Data frame or selection table (using the warbleR package's format, see \code{\link[warbleR]{selection_table}}) containing the start and end of sound events. Must contained at least the following columns: "sound.files", "selec", "start" and "end". 
 #' @param Y Data frame with the start and end of clips in the orignal sound files. Must contain the column "original.sound.files", "sound.files" (clip files), "start" and "end".
 #' @param cores Numeric. Controls whether parallel computing is applied.
 #'  It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).
@@ -9,7 +9,7 @@
 #' @return A data frame with annotations refering to the position of the detections in the original sound files.
 #' @export
 #' @name reassemble_detection
-#' @details When working with large sound files, it can be convenient to split files into smaller clips to speed up the detection process (this can be done with the function \code{\link{split_acoustic_data}}). However, this can make it difficult to interpret the results, as the detections will refer to the clips rather than the original sound files. The function will take detections made on clips created with \code{\link{split_acoustic_data}}, and format their information so they refer back to the original (unsplit) sound files.
+#' @details When working with large sound files, splitting them into smaller clips (which can be done with \code{\link{split_acoustic_data}}) can accelerate detection processing. However, this approach complicates result interpretation since detections reference the clips rather than the original files. This function reformats clip-based detections —specifically those created by split_acoustic_data()— to map them back to their original unsplit sound files.
 #'
 #' @examples {
 #' # load example data