+++ NEWS.md
@@ -1,5 +1,6 @@
 <!-- based on: https://devguide.ropensci.org/newstemplate.html#newstemplate -->
 
+
 ohun 1.0.3
 =========================